HP-2419 implemented Service for BillingRegistry and allow client code to communicate only through it
